Change syrocr getchars to accept directories

+++ syrocr.py
@@ -24,31 +24,57 @@ def command_drawboxes(args):
     im_lines.save(basename + '_lines.png', format="PNG")
 
 def command_getchars(args):
-    basename = os.path.basename(os.path.splitext(args.source_image)[0])
+    # TODO make this work both with directories and single files
+    source_img_dir = os.path.realpath(args.source_img_dir)
+    json_lines_dir = os.path.realpath(args.json_lines_dir) # TODO make optional with default
+    tables_file = os.path.realpath(args.json_tables_file) # TODO make optional with default
+    # optional settings, TODO set these in argparser
+    src_ext = '.tif'
+    json_ext = '_lines.json'
+    txtlines_ext = '_textlines.json'
 
-    with open(args.json_lines_file, 'r') as f:
-        lines = json.load(f)
+    if os.path.isfile(tables_file):
+        with open(tables_file, 'r') as f:
+            tables = json.load(f)
+        for textsize in tables:
+            for entry in tables[textsize]:
+                entry['avgim'] = AvgIm(
+                    entry['avgim']['base64_str'],
+                    entry['avgim']['baseline'],
+                    entry['avgim']['width'],
+                    entry['avgim']['height'])
+    else:
+        tables = {'normal': [], 'small': []}
 
-    with open(args.json_tables_file, 'r') as f:
-        tables = json.load(f)
-    for textsize in tables:
-        for entry in tables[textsize]:
-            entry['avgim'] = AvgIm(
-                entry['avgim']['base64_str'],
-                entry['avgim']['baseline'],
-                entry['avgim']['width'],
-                entry['avgim']['height']
-                )
-    textlines, tables = scanpage(args.source_image, lines, tables, verbose=args.verbose)
+    for i, src_img_file in enumerate(get_src_files(source_img_dir, src_ext)):
+        base = os.path.splitext(src_img_file.name)[0]
+        json_lines_file = os.path.join(json_lines_dir, base + json_ext)
+        if not os.path.isfile(json_lines_file):
+            raise FileNotFoundError('not found:', json_file)
+        if args.verbose:
+            print(f'Scanning page {i}: {src_img_file.name}')
+        textlines, tables = scanpage(src_img_file.path, json_lines_file, tables,
+                                     verbose=args.verbose)
+        # after scanning each page, save the textlines to a file
+        json_text_file = os.path.join(json_lines_dir, base + txtlines_ext)
+        with open(json_text_file, 'w') as f:
+            json.dump(textlines, f, indent=2)
 
-    with open(basename + '_textlines.json', 'w') as f:
-        json.dump(textlines, f, indent=2)
-    # TODO try converting avgim in json.dump with default serializer: https://stackoverflow.com/a/41200652
+    # after all pages have been scanned, save tables to file
+    # TODO try converting avgim in json.dump with default serializer:
+    # https://stackoverflow.com/a/41200652
     for textsize in tables:
         for entry in tables[textsize]:
             entry['avgim'] = entry['avgim'].export()
-    with open(args.json_tables_file, 'w') as f:
+    with open(tables_file, 'w') as f:
         json.dump(tables, f)
+
+def get_src_files(src_dir, src_ext='.tif'):
+    """Pair image files in src_dir with corresponding json files"""
+    with os.scandir(src_dir) as sd:
+        for dir_entry in sorted(sd, key = lambda x: x.name):
+            if dir_entry.is_file() and dir_entry.name.endswith(src_ext):
+                yield dir_entry
 
 if __name__ == "__main__":
     # initialize main argument parser
@@ -97,11 +123,11 @@ if __name__ == "__main__":
         help='increase output verbosity',
         action='store_true')
     p_getchars.add_argument(
-        'source_image',
-        help='Filename of source image')
+        'source_img_dir',
+        help='Directory with source images')
     p_getchars.add_argument(
-        'json_lines_file',
-        help='Filename of json lines file')
+        'json_lines_dir',
+        help='Directory with json lines files')
     p_getchars.add_argument(
         'json_tables_file',
         help='Filename of json tables file')
